Generify JSON Merge Task
