+++ src/main/groovy/edu/wpi/first/gradlerio/test/JavaExternalSimulationTask.groovy
@@ -12,7 +12,7 @@ import org.gradle.jvm.tasks.Jar
 class JavaExternalSimulationTask extends ExternalSimulationTask {
 
     @OutputFile
-    File outfile = new File(project.rootProject.buildDir, "debug/partial/${project.name}_java.json")
+    File outfile = new File(project.rootProject.buildDir, "${ExternalSimulationMergeTask.CONTAINER_FOLDER}/${project.name}_java.json")
 
     @TaskAction
     void create() {
